ENH: Massage the SpherePack source to compile shared libraries.
Mostly this involved factoring out subroutines defined in multiple
files into a file with common routines that I can put at the end of
the sources (Windows links can't find functions defined in a library
earlier in the link line; I don't know if this is an issue with object
files, but this avoids the issue regardless).

I also had to split libspherepack into two libraries, one for the main
functionality and the other for the visualization, which depends on
the NCAR graphics code.